Revert "Stub separating all components into connected components"
This reverts commit abd6f44848011e88f6ade25290307ee41910c591.
