+++ src/InterviewerModePrompt.js
@@ -1,11 +1,6 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import WritingSurface from './WritingSurface';
 import { animated } from 'react-spring';
-
-const mapStateToProps = state => ({
-  time: state.time,
-});
 
 const InterviewerModePrompt = (props) => {
   return (
@@ -23,7 +18,7 @@ const InterviewerModePrompt = (props) => {
   );
 }
 
-export default connect()(InterviewerModePrompt);
+export default InterviewerModePrompt;
 
 const interviewerNotePrompt = {
   border: '1px solid #000',
